Updated colors per fact
New ColorWheel class which will generate a random color for each fact.
Will convert our color string to an integer and return it from the
getColor() method.

In FunFactsActivity.java, we will get a color from the ColorWheel and
store it in the variable color as an integer.
Then, we'll set the background color of the RelativeLayout to this color
and also set the text color of the FactButton to the same color.

Also set the default view so that the background color of the
RelativeLayout is the same as the textColor of the button.